REGISTRATION-postgreSQL/requirements.txt
anyio==3.6.2
astroid==2.15.0
autoflake==2.0.1
black==23.1.0
click==8.1.3
colorama==0.4.6
dill==0.3.6
fastapi==0.93.0
flake8==6.0.0
greenlet==2.0.2
h11==0.14.0
idna==3.4
isort==5.12.0
Jinja2==3.1.2
lazy-object-proxy==1.9.0
MarkupSafe==2.1.2
mccabe==0.7.0
mypy-extensions==1.0.0
packaging==23.0
pathspec==0.11.0
platformdirs==3.1.1
psycopg2-binary==2.9.5
pycodestyle==2.10.0
pydantic==1.10.6
pyflakes==3.0.1
pylint==2.17.0
python-multipart==0.0.6
sniffio==1.3.0
starlette==0.25.0
tomlkit==0.11.6
typing_extensions==4.5.0
uvicorn==0.21.0
wrapt==1.15.0
